+++ src/Fotbiler.RuleGate.Angular/src/public-api.ts
@@ -1,5 +1,7 @@
 export * from './lib/client/rule-gate-authorization-client';
 export * from './lib/directives/rule-gate-can.directive';
+export * from './lib/directives/rule-gate-disable.directive';
 export * from './lib/guards/rule-gate-permission.guard';
 export * from './lib/guards/rule-gate-policy.guard';
 export * from './lib/models/rule-gate-authorization.models';
+export * from './lib/routing/rule-gate-route-authorization';